fix: setting events on studio components in page mode
